move logout to views & fix loginas
